requirements.txt
matplotlib
netCDF4
numpy
pyserial